btf: generate BTF information for the kernel using the toolchain
This introduces a new config option, CONFIG_BTFARCHIVE_DEDUP, which causes
the compiler to be run with -gbtf to emit information on the kernel's types
and global variables, and a tool, btfarchive, which runs before final module
link, deduplicating that information into a mmappable type archive named
vmlinux.btfa.  This archive is a CTF archive containing only BTF type kinds
plus CTF_K_CONFLICTED entries for conflicting types (invisible to consumers
except that such types return 1 when ctf_type_conflicting() is called).

You can view this archive with

objdump --ctf vmlinux.btfa

from binutils XXX.

Out-of-tree modules cannot participate in this mechanism since vmlinux.btfa
is already written: so, as happens for normal userspace binaries, the
compiler emits .BTF or .ctf sections into the modules themselves, and the
linker deduplicates them (there is as yet no attempt to dedup against the
core kernel's BTFA: this probably needs adding).  strip does not strip such
sections out, but additional care should be taken not to strip such sections
into debug RPMs (they are small enough that this should not be a problem).

For in-tree modules and the core kernel, btfarchive uses the same code (in
libctf) to do the deduplication into vmlinux.btfa as the linker does for
out-of-tree modules and userspace code compiled with -gbtf: there is
no code duplication.

Within the btfa file, the type information is divided into a shared
repository, containing all types used by more than one module and BTF for
the core kernel and all built-in modules, and separate BTF for each module
built that is not built in to the kernel

This introduces new kernel build-time dependencies on a sufficiently new
binutils (not yet upstream), a BTF-capable GCC, and libctf: if any of these
is not found, CONFIG_BTFARCHIVE_DEDUP is disabled and all this machinery has
no effect.

Source-wise, outside of various places deleting the BTF where not desired
(mostly the vDSO, the final modules, and the final kernel), there is hardly
any impact to the existing source tree: the new rules are mostly centralized
in a new scripts/Makefile.btfa, and a small utility 'btfarchive' to do the
final deduplication.